Refactor: move links replacement into utils, call it after Readability
